Enable, Disable Commands and GetPermsCommand Minor Bug fixes

+++ src/main/java/com/xuxe/falconHeavy/commands/utilities/MathCommand.java
@@ -29,7 +29,7 @@ public class MathCommand extends Command {
         } else {
             double val = Double.parseDouble(res);
             if (val - (int) val == 0) {
-                trigger.respond(val + "");
+                trigger.respond((int) val + "");
             } else {
                 trigger.respond(format.format(val));
             }
